requirements/common.txt
Django==1.10.6

pip==9.0.1
setuptools==34.3.1

celery==3.1.24
django-autocomplete-light==3.2.1
django-bower==5.2.0
django-extensions==1.7.7
django-fsm==2.5.0
django-model-utils==2.6.1
django-pipeline==1.6.11
django-reversion==2.0.8
ipdb==0.10.2
ipython==4.2.1
jupyter==1.0.0
libsass==0.12.3
Markdown==2.6.8
psycopg2==2.7
python-dateutil==2.6.0
pytz==2016.10
pyyaml==3.12
raven==5.32.0
requests==2.13.0
Werkzeug==0.11.15
